+++ magda-web-client/src/Components/Dataset/Add/withAddDatasetState.tsx
@@ -1,10 +1,11 @@
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import { RouterProps, withRouter } from "react-router-dom";
 import { connect } from "react-redux";
 
 import { loadState, State } from "./DatasetAddCommon";
 import { User } from "reducers/userManagementReducer";
 import { config } from "config";
+import { useAsync } from "react-async-hook";
 
 type Props = { initialState: State; user: User } & RouterProps;
 
@@ -17,20 +18,23 @@ function mapStateToProps(state: any) {
 
 export default <T extends Props>(Component: React.ComponentType<T>) => {
     const withAddDatasetState = (props: T) => {
-        const [state, updateData] = useState<State | undefined>(undefined);
-
         const isDisabled =
             !config.featureFlags.previewAddDataset &&
             (!props.user ||
                 props.user.id === "" ||
                 props.user.isAdmin !== true);
 
-        useEffect(() => {
-            // Once redux has finished getting a logged in user, load the state (we need to pass the current user in to populate default state)
-            loadState(props.match.params.datasetId, props.user).then(state => {
-                updateData(state);
-            });
-        }, [props.user]);
+        const [state, updateData] = useState<State | undefined>(undefined);
+        const { loading, error } = useAsync(
+            async (isDisabled, datasetId, user) => {
+                if (isDisabled || !datasetId) {
+                    return;
+                }
+                const datasetState = await loadState(datasetId, user);
+                updateData(datasetState);
+            },
+            [isDisabled, props.match.params.datasetId, props.user]
+        );
 
         if (props.isFetchingWhoAmI) {
             return <div>Loading...</div>;
@@ -45,8 +49,10 @@ export default <T extends Props>(Component: React.ComponentType<T>) => {
                     </span>
                 </div>
             );
-        } else if (!state) {
+        } else if ((!state || loading) && !error) {
             return <div>Loading...</div>;
+        } else if (error) {
+            return <div>Failed to load dataset data: {"" + error}</div>;
         } else {
             return <Component {...props} initialState={state} />;
         }