+++ CHANGES.md
@@ -23,6 +23,7 @@ General:
 -   Use helm3 in inttest:registryAuth step in CI
 -   Make Magda charts helm v3 charts
 -   Add migration document for migrating Magda release v0.0.56-RC6 or eariler to v0.0.57-0
+-   Add internal authentication provider allow user to be authenticated locally
 
 UI:
 
@@ -50,6 +51,8 @@ UI:
 -   Use IBM Plex Sans font
 -   Fixed: error notification no longer has a white background
 -   Added `add dataset review page`
+-   Make `add dataset page` save state data to registry (unless in preview mode) rather than localStorage
+-   Set flag `previewAddDataset` to `false` by default
 
 Storage: